Added functionality to find or create a node, based on a set of triples that uniquely specifies it.
This is then used to translate a hypothesis into a
new state entity in the graph, along with triples
stating the associated phenotypes, and their
comparisons with reference phenotype states.

Some work left to do here, so far I have just
implemented this for the `exhibits_phenotype`
predicate in Daniel's logic, and even then I have
not implemented the logic for additional relations
based on the phenotype (e.g. `accumulationOfChemical`).